requirements.txt
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
# SPDX-License-Identifier: MIT-0

aws-cdk-lib>=2.139.0
constructs>=10.0.0,<11.0.0
pyyaml==6.0.1
cdk-nag==2.28.33
boto3==1.34.23
# Alpha 
aws-cdk.aws-cognito-identitypool-alpha>=2.139.0a0
